backend/requirements.txt
aiosqlite==0.17.0
anyio==3.6.1
click==8.1.3
colorama==0.4.5
fastapi==0.79.0
h11==0.13.0
idna==3.3
iso8601==1.0.2
pydantic==1.9.1
pypika-tortoise==0.1.6
pytz==2022.1
sniffio==1.2.0
starlette==0.19.1
tortoise-orm==0.19.2
typing-extensions==4.3.0
uvicorn==0.18.2
pillow==9.3.0
pathlib==1.0.1
DateTime==4.7
